+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.35.3] - 2024-05-28
+### Fixed
+- Transaction page when syncing receipt
+
+## [3.35.2] - 2024-05-28
+### Changed
+- Do not fail if no block is returned from provider
+- More reliable receipt sync
+
+## [3.35.1] - 2024-05-27
+### Fixed
+- Priority for blockSync / receiptSync jobs
+
 ## [3.35.0] - 2024-05-27
 ### Added
 - Introduce Rate Limiter